backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email]==2.6.3
pydantic-settings==2.2.1
redis==5.0.1
python-dotenv==1.0.1
dashscope==1.20.0
pymilvus>=2.4.0
sentence-transformers>=2.7.0
pdfplumber>=0.10.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
